alert.go: extract from probe.go
